Move logic to hide Template selector in Shop page into Product Catalog class
